+++ packages/fhi-designsystem/src/components/fhi-grid/fhi-grid.stories.ts
@@ -1,12 +1,14 @@
-import type { Meta, StoryObj } from '@storybook/web-components-vite';
+import type { StoryObj } from '@storybook/web-components-vite';
 
-import { html } from 'lit';
+import { html } from 'lit/html.js';
 
 import { FhiGrid } from './fhi-grid.component';
 
+import { FhiStorybookMeta } from '../../../.storybook/fhi-meta';
+
 new FhiGrid();
 
-const meta: Meta<FhiGrid> = {
+const meta: FhiStorybookMeta<FhiGrid> = {
   title: 'Komponenter/Grid',
   component: 'fhi-grid',
   parameters: {
@@ -15,8 +17,32 @@ const meta: Meta<FhiGrid> = {
         excludeDecorators: true,
       },
     },
+    slotTypes: [
+      {
+        description: 'Innholdet som skal plasseres i Grid.',
+      },
+    ],
   },
-  decorators: [],
+  decorators: [
+    Story =>
+      html` <div>
+        <style>
+          fhi-grid div {
+            font-family: var(--fhi-font-family-default);
+            color: var(--fhi-color-accent-text-subtle);
+            background: var(--fhi-color-accent-background-subtle);
+            border: var(--fhi-dimension-border-width) solid
+              var(--fhi-color-accent-border-subtle);
+            padding: var(--fhi-spacing-200);
+            border-radius: var(--fhi-border-radius-150);
+            display: flex;
+            justify-content: start;
+            align-items: center;
+          }
+        </style>
+        ${Story()}
+      </div>`,
+  ],
   render: args => html`<fhi-grid gap=${args.gap}></fhi-grid>`,
   argTypes: {
     gap: {
@@ -26,9 +52,14 @@ const meta: Meta<FhiGrid> = {
       defaultValue: { summary: 'medium' },
     },
     columns: {
-      controls: 'number',
+      control: 'number',
       description: 'Bestemmer antall kolonner',
       defaultValue: { summary: '12' },
+    },
+    rows: {
+      controls: 'number',
+      description: 'Bestemmer antall rader',
+      defaultValue: { summary: '1' },
     },
   },
 };
@@ -40,24 +71,13 @@ export const Preview: Story = {
   args: {
     gap: 'medium',
     columns: 12,
+    rows: 1,
   },
   render: args => html`
-    <fhi-grid gap=${args.gap} columns=${args.columns}>
-      <div
-        style="grid-column: span 6; font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 3;  font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 3;  font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
+    <fhi-grid gap=${args.gap} columns=${args.columns} rows=${args.rows}>
+      <div style="grid-column: span 6;">Slot</div>
+      <div style="grid-column: span 3;">Slot</div>
+      <div style="grid-column: span 3;">Slot</div>
     </fhi-grid>
   `,
 };
@@ -67,44 +87,34 @@ export const LargeGrid: Story = {
   args: {
     gap: 'medium',
     columns: 12,
+    rows: 1,
   },
   render: args => html`
-    <fhi-grid gap=${args.gap} columns=${args.columns}>
-      <div
-        style="grid-column: span 4;  grid-row: span 2; font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 4;  grid-row: span 2; font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 4;  grid-row: span 2; font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 6;  font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 6;  font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 10;  font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="grid-column: span 2; font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-400); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
+    <fhi-grid gap=${args.gap} columns=${args.columns} rows=${args.rows}>
+      <div style="grid-column: span 4;">Slot</div>
+      <div style="grid-column: span 4;">Slot</div>
+      <div style="grid-column: span 4;">Slot</div>
+      <div style="grid-column: span 6;">Slot</div>
+      <div style="grid-column: span 6;">Slot</div>
+      <div style="grid-column: span 10;">Slot</div>
+      <div style="grid-column: span 2;">Slot</div>
+    </fhi-grid>
+  `,
+};
+
+export const GridWithRowsAndColumns: Story = {
+  tags: [],
+  args: {
+    gap: 'medium',
+    columns: 12,
+    rows: 8,
+  },
+  render: args => html`
+    <fhi-grid rows=${args.rows} columns=${args.columns}>
+      <div style="grid-row: span 8; grid-column: 1;">Slot</div>
+      <div style="grid-row: 1; grid-column: 2 / 13;">Slot</div>
+      <div style="grid-row: 2 / 8; grid-column: 2 / 13;">Slot</div>
+      <div style="grid-row: 8; grid-column: 2 / 13;">Slot</div>
     </fhi-grid>
   `,
 };
@@ -114,24 +124,13 @@ export const NoSpan: Story = {
   args: {
     gap: 'medium',
     columns: 3,
+    rows: 1,
   },
   render: args => html`
-    <fhi-grid gap=${args.gap} columns=${args.columns}>
-      <div
-        style="font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
-      <div
-        style="font-family: var(--fhi-font-family-default); color: var(--fhi-color-accent-text-subtle); background: var(--fhi-color-accent-background-subtle); border: var(--fhi-dimension-border-width) solid var(--fhi-color-accent-border-subtle); padding: var(--fhi-spacing-200); border-radius: var(--fhi-border-radius-150); display: flex; justify-content: start; align-items: center;"
-      >
-        Slot
-      </div>
+    <fhi-grid gap=${args.gap} columns=${args.columns} rows=${args.rows}>
+      <div>Slot</div>
+      <div>Slot</div>
+      <div>Slot</div>
     </fhi-grid>
   `,
 };